+++ Vector/ViewController/AuthenticationViewController.m
@@ -16,6 +16,8 @@
 
 #import "AuthenticationViewController.h"
 
+#import "AppDelegate.h"
+
 #import "AuthInputsView.h"
 #import "ForgotPasswordInputsView.h"
 
@@ -109,6 +111,19 @@
     MXAuthenticationSession *authSession = [MXAuthenticationSession modelFromJSON:@{@"flows":@[@{@"stages":@[kMXLoginFlowTypePassword]}]}];
     [authInputsView setAuthSession:authSession withAuthType:MXKAuthenticationTypeLogin];
     self.authInputsView = authInputsView;
+}
+
+- (void)viewWillAppear:(BOOL)animated
+{
+    [super viewWillAppear:animated];
+    
+    // Screen tracking (via Google Analytics)
+    id<GAITracker> tracker = [[GAI sharedInstance] defaultTracker];
+    if (tracker)
+    {
+        [tracker set:kGAIScreenName value:[NSString stringWithFormat:@"%@", self.class]];
+        [tracker send:[[GAIDictionaryBuilder createScreenView] build]];
+    }
 }
 
 - (void)setAuthType:(MXKAuthenticationType)authType
@@ -242,17 +257,17 @@
              __weak typeof(self) weakSelf = self;
             
             alert = [[MXKAlert alloc] initWithTitle:NSLocalizedStringFromTable(@"warning", @"Vector", nil) message:NSLocalizedStringFromTable(@"auth_missing_optional_email", @"Vector", nil) style:MXKAlertStyleAlert];
+            alert.cancelButtonIndex = [alert addActionWithTitle:[NSBundle mxk_localizedStringForKey:@"cancel"] style:MXKAlertActionStyleDefault handler:^(MXKAlert *alert) {
+                
+                __strong __typeof(weakSelf)strongSelf = weakSelf;
+                strongSelf->alert = nil;
+            }];
             [alert addActionWithTitle:[NSBundle mxk_localizedStringForKey:@"continue"] style:MXKAlertActionStyleDefault handler:^(MXKAlert *alert) {
                 
                 __strong __typeof(weakSelf)strongSelf = weakSelf;
                 strongSelf->alert = nil;
                 
                 [super onButtonPressed:sender];
-            }];
-            alert.cancelButtonIndex = [alert addActionWithTitle:[NSBundle mxk_localizedStringForKey:@"cancel"] style:MXKAlertActionStyleDefault handler:^(MXKAlert *alert) {
-                
-                __strong __typeof(weakSelf)strongSelf = weakSelf;
-                strongSelf->alert = nil;
             }];
             [alert showInViewController:self];
         }