+++ Vector/ViewController/RoomMemberDetailsViewController.m
@@ -133,6 +133,14 @@
 {
     [super viewWillAppear:animated];
     
+    // Screen tracking (via Google Analytics)
+    id<GAITracker> tracker = [[GAI sharedInstance] defaultTracker];
+    if (tracker)
+    {
+        [tracker set:kGAIScreenName value:[NSString stringWithFormat:@"%@", self.class]];
+        [tracker send:[[GAIDictionaryBuilder createScreenView] build]];
+    }
+    
     // Hide the bottom border of the navigation bar to display the expander header
     [self hideNavigationBarBorder:YES];